requirements.txt
asgiref==3.4.1
Django==3.2.7
django-admin-sortable2==1.0
Pillow==8.3.2
python-dotenv==0.19.0
pytz==2021.1
sqlparse==0.4.2
